Rename: api 통신 폴더 변경
