+++ src/app/goals/[goalId]/page.tsx
@@ -7,8 +7,8 @@ import {useEffect, useState} from 'react'
 import {useQueryClient} from '@tanstack/react-query'
 import axios from 'axios'
 
-import {goalDataApi, goalDeleteApi, goalUpdateApi} from '@/app/api/goal-api'
-import {todoDeleteApi, todoUpdateApi} from '@/app/api/todo-api'
+import {goalDataApi, goalDeleteApi, goalUpdateApi} from '@/lib/goals/api'
+import {todoDeleteApi, todoUpdateApi} from '@/lib/todos/api'
 import LoadingSpinner from '@/components/common/loading-spinner'
 import AddTodoModal from '@/components/common/modal/add-todo-modal'
 import TwoButtonModal from '@/components/common/modal/two-buttom-modal'
